refactor: move iron-resize logic to FieldMixin
